Tidy up JS, implement esbuild, remove JQuery
