+++ Dockerfile
@@ -18,9 +18,9 @@ COPY views/nunjucks ./views/nunjucks
 ## Public assets
 COPY views/assets/styles/main.css ./views/assets/styles/main.css
 COPY views/assets/styles/video-js.min.css ./views/assets/styles/video-js.min.css
+COPY views/assets/js/main-bundle.min.js ./views/assets/js/main-bundle.min.js
 COPY views/assets/fonts ./views/assets/fonts
 COPY views/assets/img ./views/assets/img
-COPY views/assets/js ./views/assets/js
 
 EXPOSE 3001
