+++ package.json
@@ -2,9 +2,9 @@
 	"name": "lib",
 	"version": "0.0.1",
 	"scripts": {
-		"build": "concurrently \"npm run compile\" \"npm run sass\"",
+		"build": "concurrently \"npm run compile\" \"npm run sass\" \"npm run js\"",
 		"compile": "tsc",
-		"dev:ui": "concurrently \"npm run watch-sass\"  \"npm run start:dev\"",
+		"dev": "nodemon src/server.ts",
 		"lint": "concurrently \"npm run eslint\" \"npm run format\"",
 		"fix-lint": "concurrently \"npm run fix-eslint\" \"npm run fix-format\"",
 		"eslint": "eslint",
@@ -12,10 +12,12 @@
 		"format": "prettier --check \"**/*.+(js|ts|json)\"",
 		"fix-format": "prettier --write \"**/*.+(js|ts|json)\"",
 		"nuke": "rm -rf dist node_modules",
+		"js": "esbuild views/assets/js/main.js --bundle --minify --sourcemap --outfile=views/assets/js/main-bundle.min.js",
+		"watch-js": "npm run js -- --watch",
 		"sass": "sass views/assets/styles/main.scss:views/assets/styles/main.css --style compressed",
 		"start": "node dist/server.js",
 		"start:build": "npm run build && npm run start",
-		"start:dev": "nodemon src/server.ts",
+		"start:dev": "concurrently \"npm run watch-sass\" \"npm run watch-js\" \"npm run dev\"",
 		"start:ts": "npm run compile && npm run start",
 		"test": "concurrently \"npm run test:ui\" \"npm run test:unit\"",
 		"test:ui": "npm run test-single:ts \"test/**/*.spec.ts\"",
@@ -71,6 +73,8 @@
 		"cheerio": "^1.0.0-rc.2",
 		"concurrently": "^7.3.0",
 		"csv": "2.0.0",
+		"esbuild": "^0.25.0",
+		"esbuild-runner": "^2.2.2",
 		"eslint": "^9.14.0",
 		"eslint-plugin-jsdoc": "^50.5.0",
 		"eslint-plugin-prefer-arrow-functions": "^3.4.1",
