+++ views/component/LeftSearchBox.html
@@ -9,7 +9,7 @@
             <fieldset>
                 <legend>
                     <h2 class="govuk-related-items__heading">
-                        <button class="filter-headings" type="button" aria-controls="accordion-learning-type" aria-expanded="true" aria-label="Learning type. Hide this section" tabindex="0">Learning type</button>
+                        <button class="filter-headings" type="button" aria-controls="accordion-learning-type" aria-expanded="true" aria-label="Learning type" tabindex="0">Learning type</button>
                     </h2>
                 </legend>
                 <div id="accordion-learning-type">
@@ -27,7 +27,7 @@
             <fieldset>
                 <legend>
                     <h2 class="govuk-related-items__heading">
-                        <button class="filter-headings" type="button" aria-controls="accordion-cost" aria-expanded="true" aria-label="Cost. Hide this section" tabindex="0">Cost</button>
+                        <button class="filter-headings" type="button" aria-controls="accordion-cost" aria-expanded="true" aria-label="Cost" tabindex="0">Cost</button>
                     </h2>
                 </legend>
                 <div id="accordion-cost">
@@ -43,7 +43,7 @@
             <fieldset>
                 <legend>
                     <h2 class="govuk-related-items__heading">
-                        <button class="filter-headings" type="button" aria-controls="accordion-your-department" aria-expanded="true" aria-label="Your department. Hide this section" tabindex="0">Your department</button>
+                        <button class="filter-headings" type="button" aria-controls="accordion-your-department" aria-expanded="true" aria-label="Your department" tabindex="0">Your department</button>
                     </h2>
                 </legend>
                 <div id="accordion-your-department">
@@ -59,7 +59,7 @@
             <fieldset>
                 <legend>
                     <h2 class="govuk-related-items__heading">
-                        <button class="filter-headings" type="button" aria-controls="accordion-other-departments" aria-expanded="true" aria-label="Other departments. Hide this section" tabindex="0">Other departments</button>
+                        <button class="filter-headings" type="button" aria-controls="accordion-other-departments" aria-expanded="true" aria-label="Other departments" tabindex="0">Other departments</button>
                     </h2>
                 </legend>
                 <div id="accordion-other-departments">
@@ -77,7 +77,7 @@
             <fieldset>
                 <legend>
                     <h2 class="govuk-related-items__heading">
-                        <button class="filter-headings" type="button" aria-controls="accordion-your-areas-of-work" aria-expanded="true" aria-label="Your areas of work. Hide this section" tabindex="0">Your areas of work</button>
+                        <button class="filter-headings" type="button" aria-controls="accordion-your-areas-of-work" aria-expanded="true" aria-label="Your areas of work" tabindex="0">Your areas of work</button>
                     </h2>
                 </legend>
                 <div id="accordion-your-areas-of-work">
@@ -95,7 +95,7 @@
             <fieldset>
                 <legend>
                     <h2 class="govuk-related-items__heading">
-                        <button class="filter-headings" type="button" aria-controls="accordion-other-areas-of-work" aria-expanded="true" aria-label="Other areas of work. Hide this section" tabindex="0">Other areas of work</button>
+                        <button class="filter-headings" type="button" aria-controls="accordion-other-areas-of-work" aria-expanded="true" aria-label="Other areas of work" tabindex="0">Other areas of work</button>
                     </h2>
                 </legend>
                 <div id="accordion-other-areas-of-work">
@@ -113,7 +113,7 @@
             <fieldset>
                 <legend>
                     <h2 class="govuk-related-items__heading">
-                        <button class="filter-headings" type="button" aria-controls="accordion-your-interests" aria-expanded="true" aria-label="Your interests. Hide this section" tabindex="0">Your interests</button>
+                        <button class="filter-headings" type="button" aria-controls="accordion-your-interests" aria-expanded="true" aria-label="Your interests" tabindex="0">Your interests</button>
                     </h2>
                 </legend>
                 <div id="accordion-your-interests">
@@ -131,7 +131,7 @@
             <fieldset>
                 <legend>
                     <h2 class="govuk-related-items__heading">
-                        <button class="filter-headings" type="button" aria-controls="accordion-other-interests" aria-expanded="true" aria-label="Other interests. Hide this section" tabindex="0">Other interests</button>
+                        <button class="filter-headings" type="button" aria-controls="accordion-other-interests" aria-expanded="true" aria-label="Other interests" tabindex="0">Other interests</button>
                     </h2>
                 </legend>
                 <div id="accordion-other-interests">
