+++ views/component/Page.html
@@ -10,7 +10,6 @@
     <title>{title} &ndash; {$i18n('proposition_name')}</title>
 	<link rel="stylesheet" href="{($req.app.locals.staticAssetRoot)}/styles/main.css">
 	<link rel="stylesheet" href="{($req.app.locals.staticAssetRoot)}/styles/video-js.min.css">
-    <script src="{($req.app.locals.staticAssetRoot)}/js/jquery.min.js"></script>
     <script async src="https://www.googletagmanager.com/gtag/js?id={($config.GOOGLE_ANALYTICS_ID)}"></script> {@html $config.GOOGLE_ANALYTICS_CODE}
 </head>
 <body>
@@ -26,14 +25,7 @@
         <slot> </slot>
     </Layout>
     <script>document.body.className = document.body.className.replace('', 'js-enabled');</script>
-    <script src="{($req.app.locals.staticAssetRoot)}/js/govuk_frontend_toolkit/govuk/show-hide-content.js"></script>
-    <script src="{($req.app.locals.staticAssetRoot)}/js/govuk_frontend_toolkit/govuk/modules.js"></script>
-    <script src="{($req.app.locals.staticAssetRoot)}/js/govuk_frontend_toolkit/govuk/govuk-template.js"></script>
-    <script type="text/javascript" src="{($req.app.locals.staticAssetRoot)}/js/accessible-autocomplete.min.js"></script>
-
-    <script src="{($req.app.locals.staticAssetRoot)}/js/main.js"></script>
-
-
+    <script src="{($req.app.locals.staticAssetRoot)}/js/main-bundle.min.js"></script>
 </body>
 </html>
