+++ views/component/PageNoNav.html
@@ -27,12 +27,7 @@
         <slot> </slot>
     </Layout>
     <script>document.body.className = document.body.className.replace('', 'js-enabled');</script>
-    <script src="/js/govuk_frontend_toolkit/govuk/show-hide-content.js"></script>
-    <script src="/js/govuk_frontend_toolkit/govuk/modules.js"></script>
-    <script src="/js/govuk_frontend_toolkit/govuk/govuk-template.js"></script>
-    <script type="text/javascript" src="/js/accessible-autocomplete.min.js"></script>
-
-    <script src="/js/main.js"></script>
+    <script src="{($req.app.locals.staticAssetRoot)}/js/main-bundle.min.js"></script>
 
 
 </body>
