+++ views/nunjucks/root/baseLayout.njk
@@ -10,7 +10,6 @@
         <title>{{ pageTitle }} - {{ i18n('proposition_name') }}</title>
         <link rel="stylesheet" href="{{ staticAssetRoot }}/styles/main.css">
         <link rel="stylesheet" href="{{ staticAssetRoot }}/styles/video-js.min.css">
-        <script src="{{ staticAssetRoot }}/js/jquery.min.js"></script>
         <script async src="https://www.googletagmanager.com/gtag/js?id={{ googleAnalyticsId }}"></script> {{ googleAnalyticsCode | safe }}
     </head>
     <body>
@@ -26,11 +25,6 @@
         {% include "./footer.njk" %}
         <div class="container container--main bottom-main"></div>
         <script>document.body.className = document.body.className.replace('', 'js-enabled');</script>
-        <script src="{{ staticAssetRoot }}/js/govuk_frontend_toolkit/govuk/show-hide-content.js"></script>
-        <script src="{{ staticAssetRoot }}/js/govuk_frontend_toolkit/govuk/modules.js"></script>
-        <script src="{{ staticAssetRoot }}/js/govuk_frontend_toolkit/govuk/govuk-template.js"></script>
-        <script type="text/javascript" src="{{ staticAssetRoot }}/js/accessible-autocomplete.min.js"></script>
-
-        <script src="{{ staticAssetRoot }}/js/main.js"></script>
+        <script src="{{ staticAssetRoot }}/js/main-bundle.min.js"></script>
     </body>
 </html>
