+++ views/page/course/display-video.html
@@ -27,11 +27,7 @@
 							<a href="http://videojs.com/html5-video-support/display-video.html" target="_blank">supports HTML5 video</a>
 						  </p>
 						</video>
-
-						<script src="{($req.app.locals.staticAssetRoot)}/js/video.min.js"></script>
 						{/if}
-						<script src="{($req.app.locals.staticAssetRoot)}/js/video.js"></script>
-
                         <input type="hidden" id="course-id" value="{course.id}">
                         <input type="hidden" id="module-id" value="{module.id}">
 						<input type="hidden" id="session-id" value="{sessionId}">
